requirements.txt
lxml~=5.3.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
langchain~=0.3.14
langchain-community~=0.3.14
langchain-core~=0.3.29
langchain-openai~=0.3.0
python-dotenv~=1.0.1
requests~=2.32.3